Remove AST module because it adds nothing but another mental indirection making it harder to follow what is going on.
